adding better print outs in reomve user/course

+++ curse.py
@@ -293,13 +293,20 @@ WHERE INSTRUCTOR.DEPT =?""", (DEPT,))
                 print("Not a valid input. Must range from 0-4 credits.")
                 creditsvalid = False
         cursor.execute("""INSERT INTO COURSE VALUES('%s', '%s', '%s', '%s','%s','%s','%s','%s','%s' );""" % (crn, Name, DEPT, courseInstructor, Time, coursedays, coursesemester, courseyear, coursecredits))
+        print("Added course: ", Name)
     #do 2nd - done
     def removeCourse(self):
         #remove based on crn 
         #print courses - display all so admin can see what courses there are
         printCourses()
         removecrn =input("What is the CRN of the course you would like to remove?")
-        cursor.execute("DELETE FROM COURSE WHERE CRN=?", (removecrn, ))
+        cursor.execute("""SELECT CRN FROM COURSE""")
+        courses = cursor.fetchall()
+        tempc = tuple(map(int, removecrn.split(',')))
+        if tempc not in courses:
+            print("No course in the system with that CRN")
+        else:
+            cursor.execute("DELETE FROM COURSE WHERE CRN=?", (removecrn, ))
         
     def addUser(self, table): #done
         #add error checking - not if already same id, id != 5 num, cant have same emails(check all user tables)
@@ -452,19 +459,37 @@ WHERE INSTRUCTOR.DEPT =?""", (DEPT,))
     def removeUser(self, table):
         if table == 's':
             print("These are the current students in the system.")
-            printStudents()
+            #printStudents()
             removeid = input("Please enter the ID of the student you would like to remove.\t")
-            cursor.execute("DELETE FROM STUDENT WHERE ID=?", (removeid, ))       
+            cursor.execute("""SELECT ID FROM STUDENT""")
+            students = cursor.fetchall()
+            temps = tuple(map(int, removeid.split(',')))
+            if temps not in students:
+                print("No Studnet in the system with that ID")
+            else:
+                cursor.execute("DELETE FROM STUDENT WHERE ID=?", (removeid, ))       
         elif table == 'i':
             print("These are the current instructors in the system.")
-            printInstructors()
+            #printInstructors()
             removeid = input("Please enter the ID of the instructor you would like to remove.\t")
-            cursor.execute("DELETE FROM INSTRUCTOR WHERE ID=?", (removeid, ))  
+            cursor.execute("""SELECT ID FROM INSTRUCTOR""")
+            insts = cursor.fetchall()
+            tempi = tuple(map(int, removeid.split(',')))
+            if tempi not in insts:
+                print("No Instructor in the system with that ID")
+            else:
+                cursor.execute("DELETE FROM INSTRUCTOR WHERE ID=?", (removeid, ))  
         elif table == 'a':
             print("These are the current admins in the system.")
-            printAdmins()
+            #printAdmins()
             removeid = input("Please enter the ID of the admin you would like to remove.\t")
-            cursor.execute("DELETE FROM ADMIN WHERE ID=?", (removeid, ))  
+            cursor.execute("""SELECT ID FROM ADMIN""")
+            admins = cursor.fetchall()
+            tempa = tuple(map(int, removeid.split(',')))
+            if tempa not in admins:
+                print("No Admin in the system with that ID")
+            else:
+                cursor.execute("DELETE FROM ADMIN WHERE ID=?", (removeid, ))  
         else:
             print("Not a valid table name")
         pass
